requirements.txt
discord.py
python-dotenv
requests
moviepy
Pillow
matplotlib
numpy